+++ changelog.md
@@ -22,7 +22,7 @@
      - init_db initialises tables in database with structure defined in a schema file;
 
 ##### Todo
-- [ ] Define structure in schema and build database
+- [x] Define structure in schema and build database
       - Table storing meal plan constraints and macro details of user
       - Table storing custom foods with their nutritional information
 - [ ] Build interface to interact with database from app